// src/app/model/UploadMaterial.ts
export class UploadMaterial {
  idTeacher?: number;
  materialType?: string;
  materialCategory?: string;
  materialName?: string;
  materialTopic?: string;
  materialFilePath?: string;
  materialFileName?: string;
  materialDescription?: string;
  idClassMaterial?: number;
  idInstitution?: number;
  metaTags?: string[];
}